move PersistFilesHandler to create_standby module
